requirements.txt
packaging>=23.0
paramiko>=3.0.0
playwright>=1.40.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.31.0
urllib3>=2.0.0
pywinrm>=0.4.3,<0.5
requests-ntlm>=1.2.0,<2
